requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.6.0
billiard==3.3.0.23
captcha==0.2.4
celery==3.1.26.post2
certifi==2018.4.16
chardet==3.0.4
Click==7.0
Django==2.1
django-celery==3.2.2
django-haystack==2.8.1
django-redis==4.10.0
django-redis-sessions==0.6.1
django-tinymce==2.7.0
djangorestframework==3.8.2
Flask==1.0.2
gevent==1.2.2
greenlet==0.4.13
idna==2.6
imagekit==1.0.6
itsdangerous==1.1.0
jieba==0.39
Jinja2==2.10
kombu==3.0.37
lxml==4.1.1
MarkupSafe==1.1.0
mutagen==1.42.0
numpy==1.14.0
olefile==0.44
paramiko==2.4.1
pika==0.11.2
Pillow==5.2.0
psycopg2==2.7.5
pycparser==2.18
pycryptodomex==3.7.2
pygame==1.9.3
PyMySQL==0.7.11
PyPDF2==1.26.0
python-alipay-sdk==1.9.1
pytz==2018.4
redis==2.10.6
requests==2.18.4
SQLAlchemy==1.2.7
urllib3==1.22
utils==0.9.0
vine==1.1.4
virtualenv==16.0.0
Werkzeug==0.14.1
Whoosh==2.7.4
xlrd==1.1.0
